Added file viewer and editor. Restructured. Added new commands. Reinforced. Added new sound effects. Changed old commands.

+++ CommandConsole.cs
@@ -11,28 +11,27 @@ using System.Buffers.Text;
 using System.Diagnostics;
 using System.Runtime.CompilerServices;
 using WMPLib;
+using Newtonsoft.Json.Linq;
+using static System.Formats.Asn1.AsnWriter;
+using System.Linq.Expressions;
 
 namespace vault_thing;
 
 public class CommandConsole()
 {
-    private const string VERSION = "a1.1.0";
+    private const string VERSION = "b1.0.0";
+
+    static JObject json = JObject.Parse(File.ReadAllText("config.json"));
 
     private string Password = "";
-    public string CurrentDirectory = File.ReadAllText("defaultdir.txt");
+    public string CurrentDirectory = json.Value<string>("defaultdir");
+    public string[] AllFilesInCurrentPath = Directory.GetFiles(json.Value<string>("defaultdir"));
+
+    public bool isMuted = json.Value<bool>("mute");
 
     public string CurrentGlobName = "";
     private string CurrentGlobPath = "";
     private List<FileWrapper> CurrentGlobContents = new();
-
-    private void PlayTimeCircuitsSFX()
-    {
-        WindowsMediaPlayer player = new()
-        {
-            URL = "time_circuits.mp3"
-        };
-        player.controls.play();
-    }
 
     #region Compression
     // nicked a good deal of this
@@ -84,7 +83,7 @@ public class CommandConsole()
     }
     #endregion
 
-    public (string, Color) ProcessCommand(string command)
+    public (string, Color) ProcessCommand(string command, (string, string) fileDetails)
     {
         string[] args = command.Split(' ');
 
@@ -97,16 +96,20 @@ public class CommandConsole()
                 "cd" => ChangeDirectory(args[1..]),
                 "ls" => ListFiles(args[1..]),
                 "mk" => MakeFile(args[1], args[2]),
-                "globulate" => MoveFileToGlob(args[1], args[2..]),
-                "extract" => ExtractFromGlob(args[1], args[2..]),
+                "dl" => DeleteFile(args[1..]),
+                "globulate" => MoveFileToGlob(args[1], args[2], args[3..]),
+                "extract" => ExtractFromGlob(args[1], args[2], args[3..]),
                 "peek" => ViewGlobContents(args[1..]),
                 "def" => SetDefaultDirectory(),
                 "open" => OpenFile(args[1]),
                 "ver" => (VERSION, Color.MistyRose),
+                "sl" => SelectFile(args[1..]),
+                "commit" => WriteToSelectedFile(fileDetails),
+                "togglemute" => ToggleMute(),
                 _ => CommandNotFoundError(args[0])
             };
-        } catch (Exception ex) {
-            return (ex.ToString(), Color.Red);
+        } catch (IndexOutOfRangeException) {
+            return ("There was an error parsing command arguments.", Color.Red);
         }
     }
 
@@ -117,11 +120,26 @@ public class CommandConsole()
         return (output, Color.MistyRose);
     }
 
+    private (string, Color) ToggleMute()
+    {
+        isMuted ^= true;
+
+        Dictionary<string, dynamic> newConfig = new()
+        {
+            { "defaultdir", CurrentDirectory },
+            { "mute", isMuted }
+        };
+
+        string json = JsonConvert.SerializeObject(newConfig, Formatting.Indented);
+        File.WriteAllText("config.json", json);
+
+        return ($"Mute is now set to '{isMuted}'.", Color.MistyRose);
+    }
+
     // the password is used as an encryption seed for each .glob
     private (string, Color) Login(string password)
     {
         Password = password;
-        PlayTimeCircuitsSFX();
         return ("Login successful.", Color.Green);
     }
 
@@ -134,14 +152,18 @@ public class CommandConsole()
             { "login", " <password> - Sets current glob encryption key to <password>." },
             { "help", " <command = *> - Displays help on <command>, by default will display all commands." },
             { "cd", " <directory> - Changes current directory scope to <directory>. <directory = '^'> retrieves parent directory." },
-            { "ls", " <type> - Lists all files / directories in current directory depending on the value of <type>. <type = -f | -d | -g> where -f is all files, -d is all directories and -g is all globs." },
-            { "mk", " <type> <name> - Creates file of <type> with name <name>. <type = -f | -g> where -f is all files and -g is all globs." },
-            { "globulate", " <file> <glob> - Moves <file> to <glob>. If <glob> is left blank, the currently selected glob is assumed." },
-            { "extract", " <file> <glob> - Extracts <file> from <glob>. If <glob> is left blank, the currently selected glob is assumed." },
+            { "ls", " <type> - Lists all files / directories in current directory depending on the value of <type>. <type = -f | -d | -g> where -f is all files, -d is all directories and -g is all globs. Use F11 and F12 to scroll through each file." },
+            { "mk", " <type> <name> - Creates file of <type> with name <name>. <type = -f | -g> where -f is for a file and -g is for a glob." },
+            { "dl", " <file> - Deletes file at <file>." },
+            { "globulate", " <file> <copy? = -m | -c> <glob> - Moves <file> to <glob>. <copy = -c> indicates to copy the file, and if <glob> is left blank, the currently selected glob is assumed." },
+            { "extract", " <file ?= *> <copy? = -m | -c> <glob> - Extracts <file> from <glob>. <copy = -c> indicates to copy the file, and if <glob> is left blank, the currently selected glob is assumed. <file = *> indicates all files should be extracted." },
             { "peek", " <glob> - Views contents of <glob>, and sets it to the currently selected glob. If <glob> is left blank, the currently selected glob is assumed."},
             { "def", " - Sets current directory to the default." },
             { "open", " <file> - Opens <file> in default application." },
-            { "ver", " - Displays current software version." }
+            { "ver", " - Displays current software version." },
+            { "sl", " <file> - Selects <file> for viewing / editing on the side. Use F11 and F12 to scroll through each file." },
+            { "commit", " - Writes changes from selected file." },
+            { "togglemute", " - Toggles mute." }
         };
 
         // checks if user is requesting help for specific command
@@ -187,22 +209,44 @@ public class CommandConsole()
             string json = Encoding.ASCII.GetString(ByteCipher.XOR(Unzip(File.ReadAllBytes(CurrentGlobPath)), Password));
 
             // get all files in selected .glob file and read file name and size
-            CurrentGlobContents = JsonConvert.DeserializeObject<List<FileWrapper>>(json);
+            try {
+                CurrentGlobContents = JsonConvert.DeserializeObject<List<FileWrapper>>(json);
+            } catch (JsonReaderException) {
+                return ("There was an error deserializing glob contents. This very likely means the password provided was incorrect.", Color.Red);
+            }
 
             string contents = "\n\n";
             foreach (FileWrapper f in CurrentGlobContents)
                 contents += "> " + f.FileName + $" ({f.Bytes.Length} B)" + '\n';
 
-            PlayTimeCircuitsSFX();
             return (contents, Color.MistyRose);
         }
         
         return ($"Error! Glob '{stringGlob}' does not exist!", Color.Red);
     }
     
-    // since the usual parsing for commands works by splitting on a space, we have to pass in an array of the rest of the arguments
-    private (string, Color) ExtractFromGlob(string fileName, string[] glob)
+    private void ExtractSingleFile(FileWrapper file, string isCopy, string globPath)
     {
+        // create a new file containing the decompressed byte data of file.Bytes and write
+        File.WriteAllBytes(CurrentDirectory + '\\' + file.FileName, Decompress(file.Bytes));
+
+        if (isCopy == "-m")
+        {
+            CurrentGlobContents.Remove(file);
+            // get the byte data of the json serialized object [JsonConvert.SerializeObject()]
+            // encrypt it [ByteCipher.XOR()]
+            // compress it [Zip()]
+            byte[] jsonBytes = Zip(ByteCipher.XOR(Encoding.ASCII.GetBytes(JsonConvert.SerializeObject(CurrentGlobContents, Formatting.Indented)), Password));
+            File.WriteAllBytes(globPath, jsonBytes);
+        }    
+    }
+
+    // since the usual parsing for commands works by splitting on a space, we have to pass in an array of the rest of the arguments
+    private (string, Color) ExtractFromGlob(string fileName, string isCopy, string[] glob)
+    {
+        if (!(isCopy == "-c" || isCopy == "-m"))
+            return ($"Argument '{isCopy}' is invalid!", Color.Red);
+
         string globPath;
         string globName;
         if (glob.Length > 0) // length greater than zero indicates user has inputted a path for the .glob
@@ -221,37 +265,51 @@ public class CommandConsole()
             // refer to above method for reading file json
             string json = Encoding.ASCII.GetString(ByteCipher.XOR(Unzip(File.ReadAllBytes(CurrentGlobPath)), Password));
 
-            // get all files in selected .glob file
-            List<FileWrapper> files = JsonConvert.DeserializeObject<List<FileWrapper>>(json);
+            try {
+                // get all files in selected .glob file
+                CurrentGlobContents = JsonConvert.DeserializeObject<List<FileWrapper>>(json);
+            } catch (JsonReaderException) {
+                return ("There was an error deserializing glob contents. This very likely means the password provided was incorrect.", Color.Red);
+            }
 
-            // acquire file by file name
-            FileWrapper file = files.FirstOrDefault(file => file.FileName == fileName);
+            if (fileName != "*")
+            {
+                // acquire file by file name
+                FileWrapper file = CurrentGlobContents.FirstOrDefault(file => file.FileName == fileName);
 
-            if (file is null)
-                return ($"File '{fileName}' does not exist in glob '{glob}'!", Color.Red);
+                if (file is null)
+                    return ($"File '{fileName}' does not exist in glob '{glob}'!", Color.Red);
 
-            // create a new file containing the decompressed byte data of file.Bytes and write
-            File.WriteAllBytes(CurrentDirectory + '\\' + fileName, Decompress(file.Bytes));
-
-            files.Remove(file);
-
-            // get the byte data of the json serialized object [JsonConvert.SerializeObject()]
-            // encrypt it [ByteCipher.XOR()]
-            // compress it [Zip()]
-            byte[] jsonBytes = Zip(ByteCipher.XOR(Encoding.ASCII.GetBytes(JsonConvert.SerializeObject(files, Formatting.Indented)), Password));
-
-            File.WriteAllBytes(globPath, jsonBytes);
-
-            return ($"Extracted '{file.FileName}' from glob '{globName}'.", Color.Green);
+                ExtractSingleFile(file, isCopy, globPath);
+            }
+            
+            else
+            {
+                // we have to define an amount here and do a manual loop because CurrentGlobContents.Count may change as files are extracted
+                if (isCopy == "-m")
+                {
+                    int count = CurrentGlobContents.Count;
+                    for (int _ = 0; _ < count; _++)
+                        ExtractSingleFile(CurrentGlobContents[0], isCopy, globPath);
+                }
+                                    
+                else
+                    foreach (FileWrapper file in CurrentGlobContents)
+                        ExtractSingleFile(file, isCopy, globPath);
+            }
+                
+            return ($"Extracted '{fileName}' from glob '{globName}'.", Color.Green);
         }
         
         return ($"Error! Glob '{globName}' does not exist!", Color.Red);
     }
 
     // since the usual parsing for commands works by splitting on a space, we have to pass in an array of the rest of the arguments
-    private (string, Color) MoveFileToGlob(string fileName, string[] glob)
+    private (string, Color) MoveFileToGlob(string fileName, string isCopy, string[] glob)
     {
-        
+        if (!(isCopy == "-c" || isCopy == "-m"))
+            return ($"Argument '{isCopy}' is invalid!", Color.Red);
+
         string? filePath = ParseDirectoryRequest(fileName);
 
         string? globPath;
@@ -272,8 +330,13 @@ public class CommandConsole()
             // refer to previous
             string json = Encoding.ASCII.GetString(ByteCipher.XOR(Unzip(File.ReadAllBytes(CurrentGlobPath)), Password));
 
-            // get all files in selected .glob file
-            List<FileWrapper> files = JsonConvert.DeserializeObject<List<FileWrapper>>(json);
+            List<FileWrapper> files;
+            try {
+                // get all files in selected .glob file
+                files = JsonConvert.DeserializeObject<List<FileWrapper>>(json);
+            } catch (JsonReaderException) {
+                return ("There was an error deserializing glob contents. This very likely means the password provided was incorrect.", Color.Red);
+            }
 
             // create a new instance of the file with compressed byte data
             FileWrapper file = new(fileName, Compress(File.ReadAllBytes(filePath)));
@@ -284,7 +347,8 @@ public class CommandConsole()
             byte[] jsonBytes = Zip(ByteCipher.XOR(Encoding.ASCII.GetBytes(JsonConvert.SerializeObject(files, Formatting.Indented)), Password));
             File.WriteAllBytes(globPath, jsonBytes);
 
-            File.Delete(filePath);
+            if (isCopy == "-m")
+                File.Delete(filePath);
 
             return ($"Moved file '{fileName}' to glob '{globName}'.", Color.Green);
         }
@@ -294,9 +358,6 @@ public class CommandConsole()
     #endregion
 
     #region Files
-    // -f refers to files
-    // -d refers to directories
-    // -g refers to globs
     private (string, Color) OpenFile(string path)
     {
         string file = ParseDirectoryRequest(path);
@@ -312,10 +373,21 @@ public class CommandConsole()
 
     private (string, Color) SetDefaultDirectory()
     {
-        File.WriteAllText(Directory.GetCurrentDirectory() + '\\' + "defaultdir.txt", CurrentDirectory);
+        Dictionary<string, dynamic> newConfig = new()
+        {
+            { "defaultdir", CurrentDirectory },
+            { "mute", isMuted }
+        };
+
+        string json = JsonConvert.SerializeObject(newConfig, Formatting.Indented);
+
+        File.WriteAllText("config.json", json);
         return ("Set current directory to default.", Color.Green);
     }
 
+    // -f refers to files
+    // -d refers to directories
+    // -g refers to globs
     private (string, Color) MakeFile(string type, string name)
     {
         string filePath = CurrentDirectory + '\\' + name;
@@ -350,7 +422,7 @@ public class CommandConsole()
 
             foreach (FileInfo file in allFiles)
             {
-                if (file.Name.EndsWith(".glob"))
+                if (file.Name.EndsWith("glob"))
                     allGlobs += "> " + file.Name + $" ({file.Length} B)" + '\n';
                 else
                     stringAllFiles += "> " + file.Name + $" ({file.Length} B)" + '\n';
@@ -387,7 +459,7 @@ public class CommandConsole()
                 string allGlobs = "";
 
                 foreach (FileInfo file in allFiles)
-                    if (file.Name.EndsWith(".glob"))
+                    if (file.Name.EndsWith("glob"))
                         allGlobs += "> " + file.Name + $" ({file.Length} B)" + '\n';
 
                 stringAllFiles += allGlobs;
@@ -399,9 +471,44 @@ public class CommandConsole()
         return (stringAllFiles, Color.MistyRose);
     }
 
+    private (string, Color) SelectFile(string[] path)
+    {
+        string? stringPath = ParseDirectoryRequest(string.Join(' ', path));
+
+        if (stringPath is null)
+            return ($"File '{string.Join(' ', path)}' does not exist!", Color.Red);
+        
+        if (stringPath.EndsWith("png") || stringPath.EndsWith("jpg") || stringPath.EndsWith("jpeg") || stringPath.EndsWith("bmp") || stringPath.EndsWith("gif"))
+            Main.SetPictureBoxImage(stringPath);
+        else
+            Main.SetViewingBoxText(File.ReadAllText(stringPath), stringPath);
+
+        return ($"Currently viewing '{stringPath}'.", Color.MistyRose);
+    }
+
+    private static (string, Color) WriteToSelectedFile((string, string) fileDetails)
+    {
+        if (fileDetails.Item2.EndsWith("png") || fileDetails.Item2.EndsWith("jpg") || fileDetails.Item2.EndsWith("jpeg") || fileDetails.Item2.EndsWith("bmp") || fileDetails.Item2.EndsWith("gif") || fileDetails.Item2.EndsWith("glob"))
+            return ("Selected file is not a text file!", Color.Red);
+
+        File.WriteAllText(fileDetails.Item2, fileDetails.Item1);
+        return ($"Wrote to file '{fileDetails.Item2}'.", Color.Green);
+    }
+
+    private (string, Color) DeleteFile(string[] path)
+    {
+        string? stringPath = ParseDirectoryRequest(string.Join(' ', path));
+
+        if (stringPath is null)
+            return ($"File '{string.Join(' ', path)}' does not exist.", Color.Red);
+
+        File.Delete(stringPath);
+
+        return ($"Deleted file '{stringPath}'.", Color.Green);
+    }
+
     private string? ParseDirectoryRequest(string requestedPath)
     {
-        requestedPath = requestedPath.Trim();
         string _currentDirectory = CurrentDirectory;
 
         // '^' indicates user is requesting to go to parent directory
@@ -427,13 +534,12 @@ public class CommandConsole()
             return ($"Directory '{stringPath}' does not exist!", Color.Red);
        
         CurrentDirectory = nextDirectory;
-        PlayTimeCircuitsSFX();
         return ($"Changed current directory scope to '{stringPath}'.", Color.Green);   
     }
     #endregion
 
     private static (string, Color) CommandNotFoundError(string command)
-    {
+    {    
         return ($"Error! Command '{command}' not found!", Color.Red);
     }
 }
